Moved `makeCffstr()` in Preview component to `useCffstr()` composition function
