+++ src/store/cff.ts
@@ -21,7 +21,6 @@ const cff = ref({
 
 export function useCff () {
     return {
-        data: computed(() => cff.value),
         abstract: computed(() => cff.value.abstract),
         commit: computed(() => cff.value.commit),
         cffVersion: computed(() => cff.value.cffVersion),
